+++ project/project_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/lxc/incus-compose/client"
+	"github.com/lxc/incus-compose/shared"
 )
 
 // fixturePath returns the path to a test fixture.
@@ -25,7 +26,7 @@ func skipLocal(t *testing.T) {
 }
 
 func skipNo73(t *testing.T, c *client.Client) {
-	if !c.Global().HasExtension(client.Incus73Extension) {
+	if !c.Global().HasExtension(shared.Incus73Extension) {
 		t.Skip("nat tests with static ip require at least incus 7.3 or 7.0.2 LTS")
 	}
 }